+++ CHANGELOG.md
@@ -4,9 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### 1.1.0 - Unreleased
 - Added `Courier.isWearableApiAvailable` method, to determine wearable API availability.
+- Added `Courier.attachMockDataApi`, `Courier.attachMockMessageApi`, and `Courier.attachMockNodeApi` methods to support mocking communication for unit testing.
 - `Bitmap` fields are now allowed in `@Deliverable` annotated classes. They are automatically sent and received as an `Asset`.
 - Now uses a singleton `GoogleApiClient` for efficiency
-- Can now use Courier's annotation in nested/inner classes
+- Fixed support for using Courier in nested/inner classes
 
 ### 1.0.3 - 2015-03-28
 - Now compatible with JDK 7+ and JRE 6+